updated nav/screen naming scheme
